google-cloud/aidemy/planner/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiofiles==24.1.0
    # via cloud-sql-python-connector
aiohappyeyeballs==2.4.6
    # via aiohttp
aiohttp==3.11.13
    # via cloud-sql-python-connector
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via httpx
asn1crypto==1.5.1
    # via scramp
attrs==25.1.0
    # via aiohttp
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
cloud-sql-python-connector==1.16.0
    # via -r requirements.in
cryptography==44.0.2
    # via cloud-sql-python-connector
dnspython==2.7.0
    # via cloud-sql-python-connector
docstring-parser==0.16
    # via google-cloud-aiplatform
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
google-api-core[grpc]==2.24.1
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-auth==2.38.0
    # via
    #   cloud-sql-python-connector
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
    #   google-genai
google-cloud-aiplatform==1.82.0
    # via langchain-google-vertexai
google-cloud-bigquery==3.30.0
    # via google-cloud-aiplatform
google-cloud-core==2.4.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-resource-manager==1.14.1
    # via google-cloud-aiplatform
google-cloud-storage==2.19.0
    # via
    #   google-cloud-aiplatform
    #   langchain-google-vertexai
google-crc32c==1.6.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-genai==1.0.0
    # via -r requirements.in
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.68.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.0
    # via google-cloud-resource-manager
grpcio==1.70.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.70.0
    # via google-api-core
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   langchain-google-vertexai
    #   langsmith
httpx-sse==0.4.0
    # via langchain-google-vertexai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain-core==0.3.40
    # via langchain-google-vertexai
langchain-google-vertexai==2.0.14
    # via -r requirements.in
langsmith==0.3.11
    # via langchain-core
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
numpy==2.2.3
    # via shapely
orjson==3.10.15
    # via langsmith
packaging==24.2
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   langchain-core
    #   langsmith
pg8000==1.31.2
    # via -r requirements.in
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.0
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
protobuf==5.29.3
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.10.6
    # via
    #   google-cloud-aiplatform
    #   google-genai
    #   langchain-core
    #   langchain-google-vertexai
    #   langsmith
pydantic-core==2.27.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via
    #   google-cloud-bigquery
    #   pg8000
pyyaml==6.0.2
    # via langchain-core
requests==2.32.3
    # via
    #   -r requirements.in
    #   cloud-sql-python-connector
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   google-genai
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9
    # via google-auth
scramp==1.4.5
    # via pg8000
shapely==2.0.7
    # via google-cloud-aiplatform
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.36
    # via -r requirements.in
tenacity==9.0.0
    # via
    #   -r requirements.in
    #   langchain-core
typing-extensions==4.12.2
    # via
    #   anyio
    #   google-cloud-aiplatform
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
urllib3==2.3.0
    # via requests
websockets==14.2
    # via google-genai
yarl==1.18.3
    # via aiohttp
zstandard==0.23.0
    # via langsmith
